chore: moved comments into a component
